requirements.txt
Flask
Flask-PyMongo
Flask-Mail
apscheduler
selenium
webdriver-manager
beautifulsoup4
lxml
flask-bcrypt
